Update to https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=a11052d98db2f2a61841f0c5ee84de4ca1b3e296, 28092021 build.
GCC VERSION: aarch64-linux-gnu-gcc (GCC) 12.0.0 20210928 (experimental)
GCC COMMIT URL: https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=a11052d98db2f2a61841f0c5ee84de4ca1b3e296
BUILDER COMMIT: f9b440467a6031d7567a6c72b1f6d8eb6bc2f4f9

+++ aarch64-linux-gnu/include/linux/idxd.h
@@ -1,4 +1,4 @@
-/* SPDX-License-Identifier: GPL-2.0 WITH Linux-syscall-note */
+/* SPDX-License-Identifier: LGPL-2.1 WITH Linux-syscall-note */
 /* Copyright(c) 2019 Intel Corporation. All rights rsvd. */
 #ifndef _USR_IDXD_H_
 #define _USR_IDXD_H_
@@ -243,8 +243,8 @@ struct dsa_completion_record {
 			uint32_t	rsvd2:8;
 		};
 
-		uint16_t	delta_rec_size;
-		uint16_t	crc_val;
+		uint32_t	delta_rec_size;
+		uint32_t	crc_val;
 
 		/* DIF check & strip */
 		struct {
